+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# [1.0.1]
+- Fixed erratic zooming in of screenshots
+- Updated Lomiri Plus, MariKit, Sapot Container and Malakiboard packages
+- New Dialer app package for the redesigned UI
+
 # [1.0.0]
 
 Initial Release